+++ R/tidy_stan.R
@@ -189,8 +189,9 @@ brms_clean <- function(out) {
     re.sd <- tidyselect::starts_with("sd_", vars = out$term)
     re.cor <- tidyselect::starts_with("cor_", vars = out$term)
     lp <- tidyselect::starts_with("lp__", vars = out$term)
+    priors <- tidyselect::starts_with("prior_", vars = out$term)
 
-    removers <- unique(c(re.sd, re.cor, lp))
+    removers <- unique(c(re.sd, re.cor, lp, priors))
 
     if (!sjmisc::is_empty(removers))
       out <- dplyr::slice(out, !! -removers)
@@ -203,8 +204,9 @@ brms_clean <- function(out) {
   re.sd <- tidyselect::starts_with("sd_", vars = colnames(out))
   re.cor <- tidyselect::starts_with("cor_", vars = colnames(out))
   lp <- tidyselect::starts_with("lp__", vars = colnames(out))
+  priors <- tidyselect::starts_with("prior_", vars = colnames(out))
 
-  removers <- unique(c(re.sd, re.cor, lp))
+  removers <- unique(c(re.sd, re.cor, lp, priors))
 
   if (!sjmisc::is_empty(removers))
     out <- dplyr::select(out, !! -removers)