+++ NEWS.md
@@ -14,6 +14,7 @@
 * Fix typo in `xtab_statistics()`.
 * Fix issue with wrong calculation of Nagelkerke's r-squared value in `r2()`.
 * Fix issue for factors with character leves in `typical_value()`, when argument `fun` for factors was set to `mode`.
+* Don't show prior-samples in `hdi()`, `tidy_stan()` etc. for _brmsfit_-objects.
 
 # sjstats 0.14.2